Perform all calculations using signed types
* Integers of the same rank get implicitly converted to unsigned types:
    warning: comparison of unsigned expression < 0 is always false [-Wtype-limits]
    if (numIns - numDel + refGenome.length < 0) {

+++ extrautils/Evolve.cpp
@@ -129,7 +129,7 @@ int main(int argc, char* argv[])
         //
         FASTASequence newSequence;
         DNALength newPos;
-        if (numIns - numDel + refGenome.length < 0) {
+        if (numIns - numDel + static_cast<int>(refGenome.length) < 0) {
             std::cout << "ERROR, the genome has been deleted to nothing." << std::endl;
             std::exit(EXIT_FAILURE);
         }
